+++ src/utils/menu.js
@@ -13,6 +13,7 @@ export function navigateMenuWithKeyboard(getItems, e) {
       index += 1;
     }
     const item = items[index];
+    console.log(items, item);
     if (item) {
       item.focus();